Refactored tests, refactored mentorship_profile.views, added test cases for re-regisering same user.
