+++ mentorship/templates/registration/logout.html
@@ -1,4 +1,4 @@
-{% extends 'imagersite/base.html' %}
+{% extends 'mentorship/base.html' %}
 {% block body %}
 You have succesfully logged out.
 <a href="{% url 'home' %}">Home</a>
