+++ mentorship_profile/templates/mentorship_profile/activate_notification.html
@@ -1,5 +1,5 @@
 {% extends 'mentorship/base.html' %}
 {% block body %}
-    <h3>Activation Almsot Complete</h3>
+    <h3>Activation Almost Complete</h3>
     <p>An email has been sent to verify your account, please use the link in the email to finish activating your account.</p>
 {% endblock %}
